Refactor FilterSearch to component for reuse on Schedules page
