+++ src/lib/components/filter-search/close-filter-button.svelte
