+++ src/lib/components/filter-search/duration-filter.svelte
