+++ src/lib/components/filter-search/number-filter.svelte
