+++ src/lib/components/filter-search/search-attribute-menu.svelte
@@ -10,9 +10,8 @@
     MenuItem,
   } from '$lib/holocene/menu';
   import { translate } from '$lib/i18n/translate';
-  import type { WorkflowFilter } from '$lib/models/workflow-filters';
-  import { workflowFilters } from '$lib/stores/filters';
-  import { sortedSearchAttributeOptions } from '$lib/stores/search-attributes';
+  import type { SearchAttributeFilter } from '$lib/models/search-attribute-filters';
+  import type { SearchAttributeOption } from '$lib/stores/search-attributes';
   import type { SearchAttributesValue } from '$lib/types/workflows';
   import {
     getFocusedElementId,
@@ -22,10 +21,13 @@
 
   import { FILTER_CONTEXT, type FilterContext } from './index.svelte';
 
+  export let filters: SearchAttributeFilter[];
+  export let options: SearchAttributeOption[];
+
   const { filter, activeQueryIndex, focusedElementId } =
     getContext<FilterContext>(FILTER_CONTEXT);
 
-  function isOptionDisabled(value: string, filters: WorkflowFilter[]) {
+  function isOptionDisabled(value: string, filters: SearchAttributeFilter[]) {
     return filters.some(
       (filter) =>
         (filter.conditional === '=' || filter.conditional === '!=') &&
@@ -41,13 +43,13 @@
 
   let searchAttributeValue = '';
   //  TODO: Add KeywordList support
-  $: options = $sortedSearchAttributeOptions.filter(
+  $: _options = options.filter(
     ({ value, type }) => !isListFilter({ attribute: value, type }),
   );
 
   $: filteredOptions = !searchAttributeValue
-    ? options
-    : options.filter((option) =>
+    ? _options
+    : _options.filter((option) =>
         option.value.toLowerCase().includes(searchAttributeValue.toLowerCase()),
       );
 </script>
@@ -56,7 +58,7 @@
   <MenuButton
     controls="search-attribute-menu"
     disabled={$activeQueryIndex !== null}
-    count={$filter.attribute ? 0 : $workflowFilters.length}
+    count={$filter.attribute ? 0 : filters.length}
     on:click={() => (searchAttributeValue = '')}
     class="text-nowrap {!!$filter.attribute && 'attribute-selected'}"
   >
@@ -88,7 +90,7 @@
     </MenuItem>
 
     {#each filteredOptions as { value, label, type }}
-      {@const disabled = isOptionDisabled(value, $workflowFilters)}
+      {@const disabled = isOptionDisabled(value, filters)}
       <MenuItem
         on:click={() => {
           handleNewQuery(value, type);
