+++ src/lib/components/workflow/dropdown-filter/workflow-datetime-filter.svelte
@@ -21,7 +21,7 @@
   import MenuDivider from '$lib/holocene/menu/menu-divider.svelte';
   import TimePicker from '$lib/holocene/time-picker.svelte';
   import { translate } from '$lib/i18n/translate';
-  import type { WorkflowFilter } from '$lib/models/workflow-filters';
+  import type { SearchAttributeFilter } from '$lib/models/search-attribute-filters';
   import { supportsAdvancedVisibility } from '$lib/stores/advanced-visibility';
   import { workflowFilters } from '$lib/stores/filters';
   import { getLocalTime } from '$lib/utilities/format-date';
@@ -80,7 +80,7 @@
     } else if (value === 'Custom') {
       custom = true;
     } else {
-      const filter: WorkflowFilter = {
+      const filter: SearchAttributeFilter = {
         attribute: timeField,
         type: 'Datetime',
         value,
@@ -154,7 +154,7 @@
         )}"`
       : `> "${formatISO(startDateWithTime)}"`;
 
-    const filter: WorkflowFilter = {
+    const filter: SearchAttributeFilter = {
       attribute: timeField,
       type: 'Datetime',
       value: query,
