+++ src/lib/services/schedule-service.ts
@@ -33,12 +33,17 @@ type PaginatedSchedulesPromise = (
 
 export const fetchPaginatedSchedules = async (
   namespace: string,
+  query: string,
   request = fetch,
 ): Promise<PaginatedSchedulesPromise> => {
   return (pageSize = 100, token = '') => {
     const route = routeForApi('schedules', { namespace });
     return requestFromAPI<ListScheduleResponse>(route, {
-      params: { maximumPageSize: String(pageSize), nextPageToken: token },
+      params: {
+        maximumPageSize: String(pageSize),
+        nextPageToken: token,
+        ...(query ? { query } : {}),
+      },
       request,
     }).then(({ schedules, nextPageToken }) => {
       return {
